Added utility function to explode dictionary into a sorted list.  This can be used when comparing two dictionaries together.
Also added additional logging support for settings file.

Also added validation support for the settings file.  The validation
only works for the keys of the dictionary/json object.  This was done
because the default settings object stored in code doesn't necessarily
have the same key:value pairs as the settings file so we can't
just compare the dictionaries/json objects.

:w
ww